requirements.txt
flask==3.0.0
google-generativeai==0.3.2
playwright==1.40.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
beautifulsoup4==4.12.2
numpy==1.26.2
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0